liana-ui: swap widget::modal && component::modal
